Changed ID from int to Long

+++ src/main/java/dao/CrudDao.java
@@ -4,10 +4,10 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CrudDao <T> {
-    Optional<T> find(Integer id);
+    Optional<T> find(long id);
     void save(T model);
     void update(T model);
-    void delete(Integer id);
+    void delete(long id);
 
     List<T> findAll();
 }
